requirements.txt
alembic==1.9.1
fastapi==0.88.0
fastapi-pagination==0.11.1
passlib[bcrypt]==1.7.4
pydantic[email]==1.10.2
psycopg2-binary==2.9.5
python-jose[cryptography]==3.3.0
sqlalchemy==1.4.45
uvicorn[standard]==0.20.0
